improve server framework

+++ src/main/java/cn/fql/fishbone/dao/PermissionDAO.java
@@ -8,5 +8,5 @@ import java.util.List;
  * Created by fuquanlin on 2016/5/23.
  */
 public interface PermissionDAO {
-    List<Permission> getPermissionsByRoleId(Integer id);
+    List<Permission> getPermissionsByRoleId(Long id);
 }
