+++ src/main/java/cn/fql/fishbone/dao/RoleDAO.java
@@ -8,5 +8,5 @@ import java.util.List;
  * Created by fuquanlin on 2016/5/23.
  */
 public interface RoleDAO {
-    List<Role> getUserRolesById(Integer id);
+    List<Role> getUserRolesById(Long id);
 }
